+++ integrations/nvidia/src/haystack_integrations/components/embedders/nvidia/document_embedder.py
@@ -4,6 +4,7 @@
 
 import os
 import warnings
+from dataclasses import replace
 from typing import Any, Optional, Union
 
 from haystack import Document, component, default_from_dict, default_to_dict, logging
@@ -49,7 +50,7 @@ class NvidiaDocumentEmbedder:
         embedding_separator: str = "\n",
         truncate: Optional[Union[EmbeddingTruncateMode, str]] = None,
         timeout: Optional[float] = None,
-    ):
+    ) -> None:
         """
         Create a NvidiaTextEmbedder component.
 
@@ -108,7 +109,7 @@ class NvidiaDocumentEmbedder:
     def class_name(cls) -> str:
         return "NvidiaDocumentEmbedder"
 
-    def default_model(self):
+    def default_model(self) -> None:
         """Set default model in local NIM mode."""
         valid_models = [
             model.id for model in self.available_models if not model.base_model or model.base_model == model.id
@@ -129,7 +130,7 @@ class NvidiaDocumentEmbedder:
             error_message = "No locally hosted model was found."
             raise ValueError(error_message)
 
-    def warm_up(self):
+    def warm_up(self) -> None:
         """
         Initializes the component.
         """
@@ -246,14 +247,12 @@ class NvidiaDocumentEmbedder:
             A dictionary with the following keys and values:
             - `documents` - List of processed Documents with embeddings.
             - `meta` - Metadata on usage statistics, etc.
-        :raises RuntimeError:
-            If the component was not initialized.
         :raises TypeError:
-            If the input is not a string.
+            If the input is not a list of Documents.
         """
         if not self._initialized:
-            msg = "The embedding model has not been loaded. Please call warm_up() before running."
-            raise RuntimeError(msg)
+            self.warm_up()
+
         elif not isinstance(documents, list) or (documents and not isinstance(documents[0], Document)):
             msg = (
                 "NvidiaDocumentEmbedder expects a list of Documents as input."
@@ -267,7 +266,9 @@ class NvidiaDocumentEmbedder:
 
         texts_to_embed = self._prepare_texts_to_embed(documents)
         embeddings, metadata = self._embed_batch(texts_to_embed, self.batch_size)
-        for doc, emb in zip(documents, embeddings):
-            doc.embedding = emb
 
-        return {"documents": documents, "meta": metadata}
+        new_documents = []
+        for doc, emb in zip(documents, embeddings):
+            new_documents.append(replace(doc, embedding=emb))
+
+        return {"documents": new_documents, "meta": metadata}